添加首帧

+++ upload-mserver/src/main/java/cn/wodesh/service/impl/UploadServiceImpl.java
@@ -2,11 +2,15 @@ package cn.wodesh.service.impl;
 
 import cn.wodesh.config.FileSource;
 import cn.wodesh.entity.Header;
+import cn.wodesh.entity.Page;
 import cn.wodesh.mapper.FileDataMapper;
 import cn.wodesh.model.FileData;
 import cn.wodesh.service.IUploadService;
 import cn.wodesh.util.*;
 import cn.wodesh.validation.FileUploadAssert;
+import com.alibaba.fastjson.JSONObject;
+import com.github.pagehelper.PageHelper;
+import com.github.pagehelper.PageInfo;
 import org.apache.tomcat.util.http.fileupload.FileUploadException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +18,7 @@ import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 import org.springframework.web.multipart.MultipartFile;
 import java.io.File;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -39,19 +44,14 @@ public class UploadServiceImpl implements IUploadService{
     @Override
     @Transactional
     public Object upload(MultipartFile multipartFile , String keyFormSize) throws Exception {
+        //判断是否选择文件
+        FileUploadAssert.MultipartFileNotNull(multipartFile);
         FileData fileData = null;
         String file_path = null;
         String keyform_path = null;
-        int keyFormHeight , keyFormWidth = 0;
-
-        if(!StringUtils.isEmpty(keyFormSize)){
-            String arr[] = keyFormSize.split("[*]");
-            keyFormHeight = Integer.parseInt(arr[0].trim());
-            keyFormWidth = Integer.parseInt(arr[1].trim());
-        }else {
-            keyFormHeight = fileSource.getKeyFormHeight();
-            keyFormWidth = fileSource.getKeyFormWidth();
-        }
+        int[] keyFormSizes = fileSource.getKeyFormSizes(fileSource.getKeyFormSize());
+        if(!StringUtils.isEmpty(keyFormSize))
+            keyFormSizes = fileSource.getKeyFormSizes(keyFormSize);
         try {
             String contentType = multipartFile.getOriginalFilename();
             String arr[] = contentType.split("[.]");
@@ -77,7 +77,7 @@ public class UploadServiceImpl implements IUploadService{
             //组装url
             String url = new StringBuffer().append(fileSource.getUrlType()).append("://")
                     .append(fileSource.getUrl()).append("/").append(filePath).append("/").append(DateUtil.currentTime(DateUtil.HHMMSS)).append("/").append(fileName).toString();
-            Map<String,String> map = imageUtil.createKeyForm(path.toString() , keyFormHeight, keyFormWidth , type);
+            Map<String,String> map = imageUtil.createKeyForm(path.toString() , keyFormSizes[0], keyFormSizes[1] , type);
             keyform_path = map.get("path");
             Header header = RequestUtil.getHeader();
             fileData = new FileData();
@@ -93,12 +93,34 @@ public class UploadServiceImpl implements IUploadService{
             fileData.setUserId(TokenUtil.getTokenParam(header.getToken()).getUserid());
             fileDataMapper.insert(fileData);
         }catch (Exception e){
-            File file = new File(file_path);
-            file.delete();
-            file = new File(keyform_path);
-            file.delete();
+            if(!StringUtils.isEmpty(file_path)){
+                File file = new File(file_path);
+                if(file != null)
+                    file.delete();
+            }
+            if(!StringUtils.isEmpty(keyform_path)){
+                File file = new File(keyform_path);
+                if(file != null)
+                    file.delete();
+            }
             throw new FileUploadException("文件上传失败");
         }
         return ResultUtil.success(FilterParam.fileDtaeFilter(fileData));
+    }
+
+    /**
+     * 分页排序查询
+     * @param page
+     * @return
+     * @throws Exception
+     */
+    @Override
+    public Object findByUserIdPage(Page<FileData> page) throws Exception {
+        Header header = RequestUtil.getHeader();
+        PageHelper.startPage(page.getCurrentPage(), page.getPageSize());
+        page.getFields(FileData.class).setUserId(TokenUtil.getTokenParam(header.getToken()).getUserid());
+        List<FileData> list = fileDataMapper.selectList(page);
+        PageInfo<FileData> pageInfo = new PageInfo<>(list);
+        return ResultUtil.success(PageUtils.pageFormat(pageInfo));
     }
 }
